RFC: SQUASH: Changes to use CPLL / GPLL for eDP on rk3399-gru
This is a squash of several proposed changes:
* drm/panel: simple: Alternate mode for Starry KR122EA0SRA 1920x1200 panel
* drm/panel: simple: Alternate mode for Sharp LQ123P1JX31 2400x1600 panel
* Documentation: bindings: Add optional pixelclock filter for analogix DP
* bridge/analogix: analogix_dp: Add optional pixelclock filter for analogix DP
* arm64: dts: rockchip: Force eDP to use CPLL / GPLL on rk3399-gru

I'm looking for feedback before I post things upstream to see if this is
totally stupid / insane or if there is a better way.

We're doing this because we want to use either CPLL (800 MHz) or
GPLL (594 MHz) for eDP on kevin and gru leaving VPLL free to change at
the whim of whatever external display is plugged in.

Previously on rk3288 Chromebooks we relied on a hack to our clock tree
to make this really work correctly (basically assuming that VOP0 was for
eDP and disabling the path from VOP0 to the dynamic PLL).  The hope is
that this is the first step on a more sane path.

NOTES:
* I'm fairly certain that these parameters should be sane for the Starry
  panel, but the Sharp panel doesn't have official min/max values in the
  datasheet I have.  It does seem to work though.
* If someone has a better suggestion, I'm all ears.

BUG=chrome-os-partner:53565
TEST=modetest; and check /sys/kernel/debug/clk/clk_summary

Original-Change-Id: I503207cc97e5807210c6a830779493f8b508ad3b
Signed-off-by: Douglas Anderson <dianders@chromium.org>

Change-Id: Ib5cd7357391409f53c5f093f0ba0159cd12b7e2d
Reviewed-on: https://chromium-review.googlesource.com/373580
Reviewed-by: Philip Chen <philipchen@chromium.org>
Tested-by: Philip Chen <philipchen@chromium.org>