FIXUP: CHROMIUM: arm64: dts: rockchip: Add rk3399-gru-kevin-r6.dts
NOTE: fixup on factory branch w/ new numbers.

We've changed the dvfs range for kevin-r6 and we expect that the next
gru (gru-r2) will have the same range.  Therefor make that range the
default for rk3399-gru and adjust all old revs for the old range (with
the assumption that support for most old revs will eventually be
dropped).

BUG=chrome-os-partner:54888
TEST=dtc

Change-Id: I62a16b1424f9e94752ba47945d3d24d26e69238d
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/368512