CHROMIUM: arm64: dts: fake wlan_module_reset_l as a regulator
It's hard to figure out a *great* place to do this, but we're supposed
to wait about 8 milliseconds after enabling the supplies before taking
the WiFi out of reset (7ms for ramp up and 1ms per Marvell's WiFi
specification). While we could try to do this with some kind of
"generic" power sequencing code (similar to what's done for SDIO in
drivers/mmc/core/pwrseq.c; or what's being proposed for some USB hubs
upstream, with [1]), there are a few problems:

(a) sequencing code like [1] is not very generic; it codifies a very
    inflexible type of reset sequence, involving only a few clocks and a
    gpio
(b) PCIe doesn't really have a nice place for handling this; all core
    bus code assumes devices have been powered and brought out of reset.
    So to use any "generic" code, we'd just have to hack it into the
    PCIe host driver itself anyway. It's not clear this will be very
    palatable upstream, and it doesn't seem to give us a lot of
    benefits, other than feeling good about using abstractions.

So, let's do the quick/dirty/simple thing of:

(1) modeling this active-low reset as an active-high regulator, whose
parent supply is the 1.8v PCIe regulator

(2) add an 8ms ramp delay to the parent regulator

(3) pull the reset low as early as possible with a "pin control hog"
(see [2]); it's currently default pull-up, and leaking current through
the powered-down chip at bootup. We also want to do this as strongly as
possible, to avoid fighting with the external pull up resistor.

Ideally, we can get the BIOS to do #3 for us too, but this is better
than nothing.

[1] https://patchwork.kernel.org/patch/9267147/
[2] https://www.kernel.org/doc/Documentation/pinctrl.txt

BUG=chrome-os-partner:56069, chrome-os-partner:54026
TEST=boot kevin to wifi

Change-Id: I3896096277e208c347644addfd4a4ca44155db33
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/371958
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>