CHROMIUM: arm64: dts: rockchip: Bump up 1.416 GHz little voltage on rk3399
We've seen some boards fail at 1.025.  Bump to 1.05 to avoid
instability.

This doesn't update the energy model.  The energy model need to be
updated soon anyway so being slightly wrong about the energy cost of the
top speed of the little cores probably won't be the end of the world
temporarily.

See CL:365270 where we'll redo this all.

BUG=chrome-os-partner:56443
TEST=Boards are more stable.

Change-Id: I6f1c85e2eac4441ff556189e75f6ad51a4600f56
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/376952
Reviewed-by: Katie Roberts-Hoffman <katierh@chromium.org>
Commit-Queue: Wonjoon Lee <woojoo.lee@samsung.com>
Tested-by: Wonjoon Lee <woojoo.lee@samsung.com>