CHROMIUM: arm64: dts: rockchip: set to CCI clock of RK3399 to 800M
This could improve the performance and reduce the memory latency.

On kevin we'll test and make sure it's stable as the issue 56389
described. Once we're sure it's good enough, this should send to
upstream, and moving from rk3399-gru.dtsi to rk3399.dtsi.

BUG=chrome-os-partner:56389
TEST=get better score of memory latency and d8.

Change-Id: I89fa6b2b0c3deec90bc6109a0535c1b7c3b71b10
Signed-off-by: Shunqian Zheng <zhengsq@rock-chips.com>
Reviewed-on: https://chromium-review.googlesource.com/377158
Reviewed-by: Katie Roberts-Hoffman <katierh@chromium.org>
Tested-by: Wonjoon Lee <woojoo.lee@samsung.com>
Commit-Queue: Wonjoon Lee <woojoo.lee@samsung.com>