FROMLIST: regulator: pwm: Add support for a fixed delay after duty cycle changes
A change of the duty cycle doesn't necessarily cause an immediate switch
to the target voltage.  On many PWM regulators there is a fixed "settle
time" (irrespective of the jump size) that we need to wait after an
upward jump.  This change introduces the device tree property
"settle-time-up-us" which allows us to specify a fixed delay after a
voltage increase.

We don't add an option of a fixed delay on the way down for now because
the way down is probably modelled best with a ramp rate, not a fixed
delay.

BUG=chrome-os-partner:54814, chrome-os-partner:56443
TEST=System is more stable

Change-Id: I4a8e2749bed388f1089b0165f5c01f0a2ab50eb9
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(am from https://patchwork.kernel.org/patch/9306303/)
Reviewed-on: https://chromium-review.googlesource.com/376878
Reviewed-on: https://chromium-review.googlesource.com/379498